refactor(oid4vp): move bean factory to dedicated class and improve cert/key loading
Extract OID4VP bean definitions from Application into Oid4VpBeanFactory,
convert CertReader to injectable singleton, and support all PEM key formats
(PKCS#8, PKCS#1, SEC1) via BouncyCastle PEMParser instead of PKCS8 only.
